worker/requirements.txt
google-generativeai==0.8.3
pika==1.3.2
redis==5.0.1
python-dotenv==1.0.0 